feat(web): add Model folders management to Settings
The model-folders UI (the ModelDirs component) lived only in the Models tab, but src/downloads/downloads.ts directs users to 'Add a model folder in Settings before downloading.' Settings had no such control, so the message was a dead end.

Extract ModelDirs into a shared component (web/src/screens/models/ModelDirs.tsx) and render it in the Settings screen (above the Hugging Face token section) as well as the Models tab. It self-applies via the existing /api/v1/modeldirs endpoints (add/remove/set-primary) through useModelDirs/useModelMutations, independent of the Save button. No backend changes.
